fix: convertir columnas STORE_STOCK a float64 antes del merge por grupos
Columnas leidas como int64 desde el Excel no aceptan NaN en pandas moderno.
Al hacer el merge con ERP (left join), SKUs sin coincidencia generan NaN en
las columnas de inventario; asignarlos a columnas int64 lanzaba el error
'Invalid value [nan] for dtype int64'. Se agrega .astype(float) en
pagina_meli_cedi_oviedo y pagina_meli_av19_bulevar_oviedo (4 bodegas ya lo tenia).

+++ MarketMasterApp.py
@@ -129,7 +129,7 @@ def pagina_meli_cedi_oviedo(feria_mode=False):
                     merged_data['PRICE'] = pd.to_numeric(merged_data['PRICE'], errors='coerce')
                     for col in merged_data.columns:
                         if "STORE_STOCK" in col:
-                            merged_data[col] = pd.to_numeric(merged_data[col], errors='coerce')
+                            merged_data[col] = pd.to_numeric(merged_data[col], errors='coerce').astype(float)
 
                     grouped = merged_data.groupby('ITEM_ID')
                     processed_groups = []
@@ -311,8 +311,8 @@ def pagina_meli_av19_bulevar_oviedo(feria_mode=False):
                     merged_data['PRICE'] = pd.to_numeric(merged_data['PRICE'], errors='coerce')
                     for col in merged_data.columns:
                         if "STORE_STOCK" in col:
-                            merged_data[col] = pd.to_numeric(merged_data[col], errors='coerce')
-                    
+                            merged_data[col] = pd.to_numeric(merged_data[col], errors='coerce').astype(float)
+
                     grouped = merged_data.groupby('ITEM_ID')
                     processed_groups = []
                     for name, group in grouped:
